+++ tests/db/schema_vouch_registry.sql
@@ -5,7 +5,7 @@ CREATE SCHEMA IF NOT EXISTS cow_protocol;
 -- These are the only fields we need from blocks.
 CREATE TABLE IF NOT EXISTS ethereum.blocks (
    "number" numeric UNIQUE,
-   time timstamptz
+   time timestamptz
 );
 
 CREATE TABLE IF NOT EXISTS cow_protocol."VouchRegister_evt_Vouch" (
@@ -13,22 +13,24 @@ CREATE TABLE IF NOT EXISTS cow_protocol."VouchRegister_evt_Vouch" (
    "bondingPool" bytea NOT NULL,
    "cowRewardTarget" bytea NOT NULL,
    sender bytea NOT NULL,
-   contract_address bytea NOT NULL,
-   evt_tx_hash bytea NOT NULL,
    evt_index int8 NOT NULL,
---    evt_block_time timestamptz NOT NULL,
-   evt_block_number int8 NOT NULL
+   evt_block_number int8 NOT NULL,
+   -- This columns are irrelevant for testing
+   contract_address bytea,
+   evt_tx_hash bytea,
+   evt_block_time timestamptz
 );
 
 CREATE TABLE IF NOT EXISTS cow_protocol."VouchRegister_evt_InvalidateVouch" (
    solver bytea NOT NULL,
    "bondingPool" bytea NOT NULL,
    sender bytea NOT NULL,
-   contract_address bytea NOT NULL,
-   evt_tx_hash bytea NOT NULL,
    evt_index int8 NOT NULL,
---    evt_block_time timestamptz NOT NULL,
-   evt_block_number int8 NOT NULL
+   evt_block_number int8 NOT NULL,
+   -- This columns are irrelevant for testing
+   contract_address bytea,
+   evt_tx_hash bytea,
+   evt_block_time timestamptz
 );
 
 -- We can either have this during setUp (uncommented) or manually during tearDown
@@ -36,4 +38,4 @@ TRUNCATE cow_protocol."VouchRegister_evt_Vouch";
 TRUNCATE cow_protocol."VouchRegister_evt_InvalidateVouch";
 TRUNCATE ethereum.blocks;
 
-INSERT INTO ethereum.blocks VALUES (0, '1970-01-01');
+INSERT INTO ethereum.blocks VALUES (0, '1970-01-01'), (1, '1971-01-01'), (2, '1972-01-01'), (3, '1973-01-01');
